+++ code/src/main/java/be/ugent/tiwi/domein/hereIncident/POINT.java
@@ -15,7 +15,7 @@ public class POINT {
     private List<DESCRIPTION_> DESCRIPTION = new ArrayList<DESCRIPTION_>();
     @SerializedName("ID")
     @Expose
-    private Integer ID;
+    private Long ID;
 
     /**
      * 
@@ -40,7 +40,7 @@ public class POINT {
      * @return
      *     The ID
      */
-    public Integer getID() {
+    public Long getID() {
         return ID;
     }
 
@@ -49,7 +49,7 @@ public class POINT {
      * @param ID
      *     The ID
      */
-    public void setID(Integer ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
