+++ code/src/main/java/be/ugent/tiwi/domein/hereIncident/POINT_.java
@@ -15,7 +15,7 @@ public class POINT_ {
     private List<DESCRIPTION____> DESCRIPTION = new ArrayList<DESCRIPTION____>();
     @SerializedName("ID")
     @Expose
-    private Integer ID;
+    private Long ID;
 
     /**
      * 
@@ -40,7 +40,7 @@ public class POINT_ {
      * @return
      *     The ID
      */
-    public Integer getID() {
+    public Long getID() {
         return ID;
     }
 
@@ -49,7 +49,7 @@ public class POINT_ {
      * @param ID
      *     The ID
      */
-    public void setID(Integer ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
