+++ code/src/main/java/be/ugent/tiwi/domein/hereIncident/ROADWAY_.java
@@ -15,7 +15,7 @@ public class ROADWAY_ {
     private List<DESCRIPTION___> DESCRIPTION = new ArrayList<DESCRIPTION___>();
     @SerializedName("ID")
     @Expose
-    private Integer ID;
+    private Long ID;
 
     /**
      * 
@@ -40,7 +40,7 @@ public class ROADWAY_ {
      * @return
      *     The ID
      */
-    public Integer getID() {
+    public Long getID() {
         return ID;
     }
 
@@ -49,7 +49,7 @@ public class ROADWAY_ {
      * @param ID
      *     The ID
      */
-    public void setID(Integer ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
